argus-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-whois==0.9.4
dnspython==2.4.2
requests==2.31.0
